requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
supabase
google-generativeai
numpy
pandas
scipy
matplotlib
httpx
psutil
scikit-learn
tensorflow
transformers
torch
